[Bug #11254] - Workstations renders gradient in bands contrary to probe values - Made output of VOI luts double instead of int. - Added IPresentationLut interface, which is the last LUT in the pipeline before rendering. Color maps are still applied separately by the renderer. Min/Max output of the IPresentationLut are settable, which made it necessary to break the API of the LutComposer, so plugin developers are forced to recognize the change. LutComposer now no longer implements IComposedLut, and you have to call GetOutputLut(minOut, maxOut); min/max out can be any range of values, but typically you would pass in the min/max output values of a grayscale display (e.g. 0, 255). Note that this change would also make it easier to support other display bit depths, since the output range can be arbitrarily defined. - Removed IDataLut implementation from AdjustableDataLut, since it doesn't really make a lot of sense. - Got rid of use of Color class in GrayscaleColorMap(s) as it is pretty inefficient to create a Color object only to convert it back to ARGB. Made sure unit tests still pass and added a unit test for the overlay version of the grayscale color map. - Fixed all the LUT unit tests, as they were all comparing VOI output against ints. Recomputed each expected value using DICOM reference function. - Fixed the probe tool so it reports VOI to 1 decimal place, just like "Raw". Probe no longer uses the "output" LUT to report VOI, since that would now actually be a p-value.
--HG--
branch : Bennu

+++ ImageViewer/Imaging/ComposableVoiLut.cs
@@ -36,9 +36,9 @@ namespace ClearCanvas.ImageViewer.Imaging
 	{
 		public abstract double MinInputValue { get; set; }
 		public abstract double MaxInputValue { get; set; }
-		public abstract int MinOutputValue { get; protected set; }
-		public abstract int MaxOutputValue { get; protected set; }
-		public abstract int this[double input] { get; }
+		public abstract double MinOutputValue { get; protected set; }
+        public abstract double MaxOutputValue { get; protected set; }
+        public abstract double this[double input] { get; }
 
 		internal override sealed double MinInputValueCore
 		{
